Separated JVMInfo and JVMVersion.
